+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - [UPDATE] default of grade in features table
 - [UPDATE] features table schema to change order to won
 - [REFACTOR] aggregation script
+- [FIX] feature values by entry time
 
 # 2.4.1 (2019/05/18)
 - [FIX] encoding error for race and horse